add mode to fail duo auth based on groups
If a user is a member of 'groups' as specified in pam_duo.conf, Duo 2FA
requires authentication. By default, if a user is NOT a member of the
groups, Duo bypasses auth by reporting PAM_SUCCESS. This makes
fallthrough auth to another pam method based on group membership
impossible.

This patch adds the group_access_fail boolean. If set, this will force
pam_duo to return PAM_SERVICE_ERR if a user is not a member of the
groups defined in the 'groups' parameter, allowing for fallthrough auth
to another pam module.

This allows Duo 2FA to be more strictly controlled, only forcing Duo 2FA
on specific users, while allowing other users to fall back to a
different style of 2FA based on their group membership.